Move test to its own file
bun-serve-html.test.ts has two pre-existing {hmr:false} timeouts under
debug+ASAN (React bundling takes ~2.6s per request there) that are unrelated
to this change.
